Update package.json to npm manual specifications
This commit makes two main changes. First, it removes all
explicit references to node_modules/.bin, modules' bin/, and
'npm bin', which were not necessary because npm implicitly appends
its bin to the end of the PATH when running scripts. A correlary
of this change was installing jasmine-core, which is no longer
bundled with karma and was necessary when the reference to the
karma executable changed.

The second change was to update the location of the entry point when
requiring the module in node.js. It was pointing to lib/adapters/node,
which was not written to in the build process. It was changed to
point to src/adapters/node, which is still written in Javascript and
is, at least for the moment, a working entrypoint when requiring
the module in node.